The .gd file
The “.gd” file, “myfirstlevel.gd” in my situation, will create the definition for our DGO file, aka our full level file. In short, we’ll have to put everything that needs to be included inside of our level here.
For now, we’ll just have to change a few names as such: