The .jsonc file
The “.jsonc” file, “my-first-level.jscon” in my situation, has multiple purposes:
Telling the compiler which 3D file will be used to create the geometry of our level
Setting up some automatic wall detection
Spawning all the actors in our level

We’ll keep the current automatic wall detection settings for now. More about those in the Advanced collisions part of this guide.

The actors are all the interactive things in the level that aren’t geometry, so every collectable, moving platforms, enemy but also particles, sound effects, etc. More about those in the Common actors, Advanced actors, and Particle systems parts of this guide.

For now we’ll just edit the names so the compiler use the right files as such