Setting up to build the level in game.gp
We’ll skip over the .glb for now as there’s some other code we need to add and edit for the decompiler to even use our new files.

Once again we can copy the already existing test-zone and edit the names to ours.

Location: “data\goal_src\jak1”
Line: 1621