Creating the checkpoint in the level-info.gc
We’re almost done to be able to get to our level in the game. The last thing we’ll need is to create a continue-point to be able to load and spawn into it.

Once again we can copy the already existing test-zone and edit it to fit our level.

Location: “data\goal_src\jak1\engine\level\level-info.gc”
Line: 1975