Finding the continue-point’s position
To know exactly where we’ll set Jak’s position, we’ll use Blender. We’ll see later in “Checkpoints and load boundaries” how to get all the information of jak and the camera.

Back into Blender, you can do a “Shift + Right click” anywhere on your mesh to set the cursor’s position on that face. We could already take the position of the cursor but I find it really useful to create empties for the position of every “continue-point” as well as actors.

Press “Shift + A” and select “Empty > any option”. I usually use Arrows for “continue-point”

The empty will serve as a point in space we can always look back at and change. Name it the same as the name of the continue-point in “level-info.gc” so you can easily know what it is and find it through the search.

Now that you have an empty set for your continue-point, you can look at its position to copy them over to “level-info.gc”. Don’t forget the axes conversion.