Adding our level to the list
If you’d try to go in the game right now, you would still not be able to find your continue-point in the list. We first need to tell the compiler to add it. To do so, go to the bottom of the “level-info.gc” file.
Once again, you can copy the line for test-zone and edit the name.

THIS ALSO FIX THE OCEAN AND INDEX TO AVOID THE CRASH