Changing the debug continue-point - If you’re using the mod-base
The debug continue-point can be changed, doing so can save you a lot of time if you have to do a lot of iteration and always need to open the debug menu to load in your level.
If you’re using the mod-base
The mod-base has an extra file for mod related settings. This file also contains an overwrite for the starting debug continue-point.
Navigate to: “data\goal_src\jak1\engine\mods\mods-settings.gc”
Line: 40

Then simply edit the continue-point to the one in your level.