Changing the debug continue-point - If you’re not using the mod-base
The debug continue-point can be changed, doing so can save you a lot of time if you have to do a lot of iteration and always need to open the debug menu to load in your level.

If you’re not using the mod-base
Even if you’re not using the mod-base, you’ll still be able to change this which I’d still suggest doing. The compile time whenever you want to change it will be a lot longer but it’s probably still going to be a huge time save overall.

Navigate to: “data\goal_src\jak1\engine\game\game-info.gc”
Line: 219
Find the “*debug segment” continue-point, which should be “village1-hut” by default and replace it with your level’s continue-point.