The “fuel-cell” etype
Power-cells also don’t have too much going on either. The only thing special you need to take care of is the task which will be required if you want your cell to actually count. This part will however be explained in the “Tasks and menu” part of the guide.
For now we’ll just keep it at 0:

Same as for Orbs, you should move them up by around 2.0 meters up from the ground for it to look normal.