“alt-actor” for blocked vents
By adding an “alt-actor” in the lump, you can set the vent to only be active if another actor has been broken. You can set a variety of different actors for it but it has to be one that is broken, not just collected like an orb or cell.
Any type of crate will do fine and be what I’ll use in this example as they’re the only common actor that you can break.

First you’ll need an actor to break, I’ll use a bucket here:
Then an ecovent with an alt-actor set to this breakable actor we just made:

In game you’ll just see the bucket:
Then after breaking it, the vent will start working: